resolve todos
